[refactor] Change DIR
related to #31
